pyproject.toml
[project]
name = "mymonitor"
version = "0.1.0"
description = "A build process memory monitor."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "plotly",                  # For interactive plotting
    "polars-lts-cpu[pyarrow]", # For data analysis and manipulation, pyarrow for better CSV/Parquet performance
    "psutil",                  # For system and process utilities
    "toml",                    # For reading project configuration files
]

[project.scripts]
mymonitor = "mymonitor.main:main_cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    # You can add other dev tools here like linters (flake8, ruff), formatters (black)
]
export = [ # New group for optional export functionality
    "kaleido",
]
